+++ .gitmodules
@@ -1,0 +1,3 @@
+[submodule "ipyvuelink"]
+	path = ipyvuelink
+	url = https://github.com/mariobuikhuizen/ipyvuelink.git
