+++ package.json
@@ -10,21 +10,23 @@
   "dependencies": {
     "core-js": "^3.1.4",
     "vue": "^2.6.10",
-    "vuetify": "^1.5.16"
+    "vuetify": "^1.5.18"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^3.9.2",
-    "@vue/cli-plugin-eslint": "^3.9.2",
-    "@vue/cli-service": "^3.9.2",
-    "@vue/eslint-config-prettier": "^4.0.1",
-    "babel-eslint": "^10.0.2",
+    "@vue/cli-plugin-babel": "^3.11.0",
+    "@vue/cli-plugin-eslint": "^3.11.0",
+    "@vue/cli-service": "^3.11.0",
+    "@vue/eslint-config-prettier": "^5.0.0",
+    "babel-eslint": "^10.0.3",
     "eslint": "^5.16.0",
+    "eslint-plugin-prettier": "^3.1.0",
     "eslint-plugin-vue": "^5.2.3",
-    "stylus": "^0.54.5",
+    "prettier": "^1.18.2",
+    "stylus": "^0.54.7",
     "stylus-loader": "^3.0.2",
-    "vue-cli-plugin-vuetify": "^0.5.0",
+    "vue-cli-plugin-vuetify": "^0.6.3",
     "vue-template-compiler": "^2.6.10",
     "vuetify-loader": "^1.2.2",
-    "webpack": "^4.35.3"
+    "webpack": "^4.40.2"
   }
 }
